Update DustManager so "it just works"
This adds a (conceptual) variable that tracks the amount of tier+1 dust that was present when we started making lumps. By always keeping to either the safety buffer or 10 below that amount (whichever is smaller), we avoid destroying the whole stack while we're waiting to craft. This makes everything "just work": You always have a buffer, but you're also crafting lumps even if the tier your uptiering to is below the safety amount.