Large re-write of DustManager
Make safety buffers dynamic, based on a relative value curve which is configured with a string.

This is both much more configurable, doesn't change the program size when you change the config, and it means that it scales well no matter how much ore you get.
The downside is that it is *really* complicated. There is *so* much garbage packed into every one of these lines.