Update D0S.DustManager
Decrease lines, but further increase complexity: Remove the lines that dealt exclusively with crushing (and the variable that tracked ore_tier),
by folding it into the main loop and always crushing the highest available ore. This makes the loop tighter and faster,
at the expense of a more complicated (and larger-to-import) script.

Also revise the default multiplier settings, based on a thorough survey of *all* high-tier producers, including regular gem-prods.
The new settings give much more low-tier dust without significantly reducing high-tier dust, and should do away with the issue
of not having enough low-tier dust to make something.