Introduce helper macros for using craft_queue.
The macros cut out a great deal of boilerplate, making the code much
easier to understand. They also revealed a couple of instances where
there were inconsistencies in how things were written; those are now
fixed.

If the naming scheme is ever changed (say, reducing the number of
characters to cut down on space), it can now be done easily instead of
it being a huge change.

No changes to README because there are no (relevant) changes to the
generated code.