Replace all the craft* scripts with a single loop.
By continuing to make use of the factory_constants structure, (almost)
the entirety of the crafting code can be replaced by a single loop. With
attention to optimization, running the loop takes slightly fewer cycles
than the hardcoded way, because it skips crafting when it's not needed.

This drops the bundle import size to 212Kb, and the number of scripts to 19.

Also tweaked the README to account for a previous change.