Change "launch factory craft" to use strings for items.
Instead of setting global variables, this changes target_tier to be a
string. This string is checked for validity and there are actual error
messages returned for bad items or the tier being out of range.

The default item has been changed to be invalid, to better show off what
needs to be changed in "launch factory craft".