Change numbering scheme on tiers to t-1
This makes some code a little more complicated, but other code a lot
simpler. It also eases certain restrictions on item ordering.

As part of this refactor, the data string + lookup was split into two,
one line to assign it to a variable and another to do the lookup. It was
necessary because of trickle-down effects from restructuring, but I
think this ends up being easier to follow.