Update D0S.DustManager
Fix a bug where the dust manager wouldn't uptier dust if there was just 1 of it present. It will now create an ore-lump if there is 10 of the previous tier dust.
Added crush_time, it's used to determine how much time to crush, such that the crusher is occupied for crush_time ammount of time. This still keeps the ore_buffer into account.
Added Editor_actions lib to make ore_state easier to understand.
Added budget system to script so it goes through every tier in 1 frame